requirements.txt
tensorflow==2.14.0
torch==2.1.0
torchvision==0.16.0
torchaudio==0.16.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.74
Pillow==10.0.0
scikit-learn==1.3.0
jupyter==1.0.0
tensorboard==2.14.0
albumentations==1.3.0
Flask==2.3.3
fastapi==0.103.0
uvicorn==0.23.2
requests==2.31.0
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
scipy==1.11.2
